+++ faircrawler/http/__init__.py
@@ -1,3 +1,13 @@
 from .client import HTTPClient, HTTPResponse
+from .url import URL
+from .exceptions import FairCrawlerError, TooManyRedirectsError, SSLVerificationError, ConnectionError
 
-__all__ = ["HTTPClient", "HTTPResponse"]
+__all__ = [
+    "HTTPClient",
+    "HTTPResponse",
+    "URL",
+    "FairCrawlerError",
+    "TooManyRedirectsError",
+    "SSLVerificationError",
+    "ConnectionError",
+]